scrapper/requirements.txt
pandas
selenium
seleniumbase
python-dotenv